pyproject.toml
# -----------------------------
# Project metadata
# -----------------------------
[project]
name = "langchain-ai-agent"
version = "0.1.0"
description = "A production-ready LangChain agent for document automation, routing, and structured output generation."
authors = [
  { name = "Ramin" }
]
readme = "README.md"
requires-python = ">=3.9"

# -----------------------------
# Backend runtime dependencies
# -----------------------------
dependencies = [
  # Core LLM pipelines
  "langchain>=0.1.14",
  "langchain-core>=0.1.34",
  "langchain-community>=0.0.30",
  "langchain-openai>=0.1.6",
  "langgraph==0.3.23",
  
  # Embedding + vector store
  "sentence-transformers>=2.6.1",
  "faiss-cpu>=1.7.4",
  "transformers>=4.38.2",
  "operators>=1.0.1",

  # Document ingestion
  "unstructured[all-docs]>=0.11.4",
  "pdfminer.six>=20221105",
  "python-docx>=1.1.0",
  "python-dotenv>=1.0.0",

  # Crawling
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.3",
  "tldextract>=5.1.2",

  # Utilities
  "PyYAML>=6.0.1",
  "pydantic>=2.6.4",
  "tqdm>=4.66.2",
  "httpx>=0.27.0",
  "loguru>=0.7.2",

  # Serving / API
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",

  # DeepSeek OpenAI-compatible API
  "openai>=1.14.2"
]

# -----------------------------
# Dev/test/formatting extras
# -----------------------------
[project.optional-dependencies]
dev = [
  "black",
  "flake8",
  "pytest",
  "pytest-cov",
  "mypy"
]

# -----------------------------
# Package discovery
# -----------------------------
[tool.setuptools.packages.find]
where = ["langchain_ai_agent"]

# -----------------------------
# Build system backend
# -----------------------------
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"
